refactor(files): extract automation settings hook
